requirements.txt
django==3.1
djangorestframework==3.12
celery==5.2.7
redis==3.5.3
flower==2.0.0
django-celery-results==2.0.0
django-celery-beat==2.2.0
pyyaml==6.0.1
jinja2==3.1.4
pymysql==1.1.0
djangorestframework-simplejwt==4.6.0
